Allow users to get/list/watch/create/delete services
As of dask/dask-kubernetes#162 the `KubeCluster` manager will need to be able to create kubernetes services. This is because the scheduler will no longer be be local to the `KubeCluster` object but instead a pod within the cluster. In order to route traffic from the local session to the scheduler a service needs to be created.