Remove deprecated chartpress logic and bug workarounds
`--image-prefix` argument, `imagePrefix` configuration of chartpress in
travis.yml, and `images` configuration in travis.yml can now be resolved
by omission. They were introduced because past versions of chartpress
required them to be defined.

The `--commit-range` argument is already in 0.3.2 where we bump
chartpress from deprecated and ignored because the information is no
longer needed for chartpress logic.